+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.47.0
 
 * Fixed MirrorMaker 2 client rack init container override being ignored.
+* Kafka nodes are now configured with PEM certificates instead of P12/JKS for keystore and truststore.
 
 ## 0.46.0